Filter order and uniqueness
Filters should be ordered by:
 - default filters from outer configuration first (and not from configurations outside the scope of where the controller is configured)
 - package level filter annotattions
 - class level filter annotattions
 - method level filter annotattions

Filters should be overriding eachother if they have the same custom filter annotation and the same filter implementation.

Also, it should not matter in which order the methods in the configuration lambda is called. controllers should respect all configurations within the lambda, and hence should be evaluated last